+++ apps/v4/app/(internal)/sink/components/avatar-demo.tsx
@@ -31,7 +31,10 @@ export function AvatarDemo() {
           <AvatarFallback>CN</AvatarFallback>
         </Avatar>
         <Avatar>
-          <AvatarImage src="https://github.com/leerob.png" alt="@leerob" />
+          <AvatarImage
+            src="https://github.com/maxleiter.png"
+            alt="@maxleiter"
+          />
           <AvatarFallback>LR</AvatarFallback>
         </Avatar>
         <Avatar>
@@ -48,7 +51,10 @@ export function AvatarDemo() {
           <AvatarFallback>CN</AvatarFallback>
         </Avatar>
         <Avatar>
-          <AvatarImage src="https://github.com/leerob.png" alt="@leerob" />
+          <AvatarImage
+            src="https://github.com/maxleiter.png"
+            alt="@maxleiter"
+          />
           <AvatarFallback>LR</AvatarFallback>
         </Avatar>
         <Avatar>
@@ -65,7 +71,10 @@ export function AvatarDemo() {
           <AvatarFallback>CN</AvatarFallback>
         </Avatar>
         <Avatar>
-          <AvatarImage src="https://github.com/leerob.png" alt="@leerob" />
+          <AvatarImage
+            src="https://github.com/maxleiter.png"
+            alt="@maxleiter"
+          />
           <AvatarFallback>LR</AvatarFallback>
         </Avatar>
         <Avatar>